refactor(launchpad): per-phase register(app) CLI registration (#7)
zilliz_ops.py held the full Typer command tree (357 lines) for all six
phase subcommands, coupling unrelated phases. Move each command body
(options, validation, result printing) into its phase module's
register(app) entrypoint and shrink zilliz_ops.py to a thin shell that
iterates an ordered phase registry.

- Add lib/cli.py: shared fail() helper (leaf module, no import cycle)
- Each lib/phases/<phase>.py exports register(app)
- zilliz_ops.py: 357 -> 51 lines; `from zilliz_ops import app` preserved
- Add tests/test_zilliz_ops_shell.py: 80-line budget guard,
  registration check, single-file extensibility proof

No CLI behavior change: --help byte-identical across app + 6 subcommands;
suite green (284 passed, same 2 pre-existing pypdf failures), mypy strict
and ruff clean.

Adapts issue #7's argparse add_subparser() to the codebase's Typer stack
via register(app) (equivalent intent, no behavioral risk).

Closes #7
